cleanup and image function definitions
